# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: string;
  username: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface CreateUserBody {
  username: string;
  isAdmin?: boolean;
}

export interface Product {
  id: string;
  name: string;
  description?: string | null;
  coverImage?: string | null;
  stock: number;
  priceBuy: number;
  priceSell: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description?: string | null;
  coverImage?: string | null;
  /** @minimum 0 */
  stock: number;
  /** @minimum 0 */
  priceBuy: number;
  /** @minimum 0 */
  priceSell: number;
}

export interface UpdateProductBody {
  name?: string;
  description?: string | null;
  coverImage?: string | null;
  /** @minimum 0 */
  stock?: number;
  /** @minimum 0 */
  priceBuy?: number;
  /** @minimum 0 */
  priceSell?: number;
}

export interface Sale {
  id: string;
  productId: string;
  productName: string;
  userId: string;
  username: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  buyerName?: string | null;
  createdAt: string;
}

export interface CreateSaleBody {
  productId: string;
  userId: string;
  /** @minimum 1 */
  quantity: number;
  buyerName?: string | null;
}

export interface SalesSummary {
  totalRevenue: number;
  totalUnits: number;
  totalSales: number;
  lowStockCount: number;
}

export interface PresenterSales {
  userId: string;
  username: string;
  totalRevenue: number;
  totalUnits: number;
  totalSales: number;
}

export interface TopBook {
  productId: string;
  name: string;
  coverImage?: string | null;
  totalUnits: number;
  totalRevenue: number;
}

export type LoginBody = {
  username: string;
};

export type ListSalesParams = {
  userId?: string;
};