# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Username-only login
 */
export const LoginBody = zod.object({
  username: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary List all users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  username: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a user (presenter or admin)
 */
export const createUserBodyIsAdminDefault = false;

export const CreateUserBody = zod.object({
  username: zod.string(),
  isAdmin: zod.boolean().default(createUserBodyIsAdminDefault),
});

export const CreateUserResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  isAdmin: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a user
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteUserResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  coverImage: zod.string().nullish(),
  stock: zod.number(),
  priceBuy: zod.number(),
  priceSell: zod.number(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product
 */
export const createProductBodyStockMin = 0;

export const createProductBodyPriceBuyMin = 0;

export const createProductBodyPriceSellMin = 0;

export const CreateProductBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  coverImage: zod.string().nullish(),
  stock: zod.number().min(createProductBodyStockMin),
  priceBuy: zod.number().min(createProductBodyPriceBuyMin),
  priceSell: zod.number().min(createProductBodyPriceSellMin),
});

export const CreateProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  coverImage: zod.string().nullish(),
  stock: zod.number(),
  priceBuy: zod.number(),
  priceSell: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.string(),
});

export const updateProductBodyStockMin = 0;

export const updateProductBodyPriceBuyMin = 0;

export const updateProductBodyPriceSellMin = 0;

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  coverImage: zod.string().nullish(),
  stock: zod.number().min(updateProductBodyStockMin).optional(),
  priceBuy: zod.number().min(updateProductBodyPriceBuyMin).optional(),
  priceSell: zod.number().min(updateProductBodyPriceSellMin).optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  coverImage: zod.string().nullish(),
  stock: zod.number(),
  priceBuy: zod.number(),
  priceSell: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteProductResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary List sales (optionally filtered by user)
 */
export const ListSalesQueryParams = zod.object({
  userId: zod.coerce.string().optional(),
});

export const ListSalesResponseItem = zod.object({
  id: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  quantity: zod.number(),
  unitPrice: zod.number(),
  totalPrice: zod.number(),
  buyerName: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListSalesResponse = zod.array(ListSalesResponseItem);

/**
 * @summary Record a sale (decrements stock)
 */

export const CreateSaleBody = zod.object({
  productId: zod.string(),
  userId: zod.string(),
  quantity: zod.number().min(1),
  buyerName: zod.string().nullish(),
});

export const CreateSaleResponse = zod.object({
  id: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  userId: zod.string(),
  username: zod.string(),
  quantity: zod.number(),
  unitPrice: zod.number(),
  totalPrice: zod.number(),
  buyerName: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Overall sales summary (revenue, units, low stock)
 */
export const GetSalesSummaryResponse = zod.object({
  totalRevenue: zod.number(),
  totalUnits: zod.number(),
  totalSales: zod.number(),
  lowStockCount: zod.number(),
});

/**
 * @summary Sales totals grouped by presenter
 */
export const GetSalesByPresenterResponseItem = zod.object({
  userId: zod.string(),
  username: zod.string(),
  totalRevenue: zod.number(),
  totalUnits: zod.number(),
  totalSales: zod.number(),
});
export const GetSalesByPresenterResponse = zod.array(
  GetSalesByPresenterResponseItem,
);

/**
 * @summary Best selling books
 */
export const GetTopBooksResponseItem = zod.object({
  productId: zod.string(),
  name: zod.string(),
  coverImage: zod.string().nullish(),
  totalUnits: zod.number(),
  totalRevenue: zod.number(),
});
export const GetTopBooksResponse = zod.array(GetTopBooksResponseItem);