# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserBody {
  username: string;
  isAdmin?: boolean;
}