# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  description?: string | null;
  coverImage?: string | null;
  stock: number;
  priceBuy: number;
  priceSell: number;
  createdAt: string;
}