# lib/api-zod/src/generated/types/topBook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface TopBook {
  productId: string;
  name: string;
  coverImage?: string | null;
  totalUnits: number;
  totalRevenue: number;
}