# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Fair Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductBody {
  name?: string;
  description?: string | null;
  coverImage?: string | null;
  /** @minimum 0 */
  stock?: number;
  /** @minimum 0 */
  priceBuy?: number;
  /** @minimum 0 */
  priceSell?: number;
}